250416/09-02-DRF/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
Django==4.2.20
djangorestframework==3.16.0
drf-spectacular==0.28.0
exceptiongroup==1.2.2
executing==2.2.0
importlib_metadata==8.6.1
inflection==0.5.1
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
rpds-py==0.24.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
wcwidth==0.2.13
zipp==3.21.0
